requirements.txt
ultralytics
opencv-python
mss
numpy
pillow
keyboard
